% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRMLog.R
\name{crm_log}
\alias{crm_log}
\title{CRM日志查询}
\usage{
crm_log(FNumber, FOccurrenceTime_min, FOccurrenceTime_max, FCompanyName)
}
\arguments{
\item{FNumber}{编码}

\item{FOccurrenceTime_min}{最小日期}

\item{FOccurrenceTime_max}{最小日期}

\item{FCompanyName}{公司代码}
}
\value{
返回值
}
\description{
CRM日志查询
}
\examples{
crm_log()
}
