% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRMLog.R
\name{ecs_log}
\alias{ecs_log}
\title{ECS日志查询}
\usage{
ecs_log(FNumber, FOccurrenceTime_min, FOccurrenceTime_max, FCompanyName)
}
\arguments{
\item{FNumber}{编码}

\item{FOccurrenceTime_min}{最小日期}

\item{FOccurrenceTime_max}{最小日期}

\item{FCompanyName}{公司代码}
}
\value{
返回值
}
\description{
ECS日志查询
}
\examples{
ecs_log()
}
